+++ routes/cartRouter.js
@@ -5,6 +5,8 @@ const { loginRequired } = require("../utils/auth");
 const router = express.Router();
 
 router.post("", loginRequired, cartController.createCart);
+router.get("", cartController.getCartList);
+router.delete("", cartController.deleteCart);
 
 module.exports = {
   router,
