+++ routes/plantRouter.js
@@ -3,9 +3,9 @@ const plantController = require("../controllers/plantController");
 
 const router = express.Router();
 
+router.post("/main", plantController.plantsList);
 router.get("/:plantId", plantController.getPlantInfo);
 router.get("", plantController.plantListFilterData);
-router.get("/main", plantController.plantsList);
 
 module.exports = {
   router,
