+++ services/cartService.js
@@ -20,6 +20,16 @@ const createCart = async (
   );
 };
 
+const getCartList = async (userId) => {
+  return cartsDao.getCartList(userId);
+};
+
+const deleteCart = async (cartId) => {
+  return cartsDao.deleteCart(cartId);
+};
+
 module.exports = {
   createCart,
+  getCartList,
+  deleteCart,
 };
